Feat: Extract Validation into it's own pipe
