+++ demo/api/src/importer/pipes/parsers/data-transformer.pipe.ts
@@ -1,15 +1,9 @@
 import { LoggerService } from "@nestjs/common";
 import { ImporterInputClass } from "@src/importer/importer-input.type";
 import { plainToInstance } from "class-transformer";
-import { validate } from "class-validator";
 import { Transform as StreamTransform, TransformCallback } from "stream";
 
-import { ImporterPipe, PipeData, ValidationError } from "../importer-pipe.type";
-
-type PipeDataAndErrors = {
-    data: PipeData;
-    errors: ValidationError[];
-};
+import { ImporterPipe } from "../importer-pipe.type";
 
 export class DataTransformerPipe implements ImporterPipe {
     constructor(private readonly inputClass: ImporterInputClass) {}
@@ -29,16 +23,8 @@ export class DataTransformer extends StreamTransform {
 
     async _transform(inputData: ParserPipeData, encoding: BufferEncoding, callback: TransformCallback) {
         try {
-            const { data, errors } = await this.convertToInstanceAndValidate(inputData);
-
-            if (errors.length > 0) {
-                errors.forEach((error) => {
-                    this.logger.error(`Validation Errors: ${JSON.stringify(error)}`);
-                });
-                this.push(null);
-                return callback(new Error("Too many validation errors"));
-            }
-            this.push(data);
+            const instance = plainToInstance(this.inputClass, inputData) as Record<string, unknown>;
+            this.push(instance);
             // eslint-disable-next-line  @typescript-eslint/no-explicit-any
         } catch (err: any) {
             this.logger.error(`Error transforming Data: ${err}`);
@@ -47,19 +33,5 @@ export class DataTransformer extends StreamTransform {
         }
 
         callback();
-    }
-
-    async convertToInstanceAndValidate(data: PipeData): Promise<PipeDataAndErrors> {
-        const instance = plainToInstance(this.inputClass, data) as Record<string, unknown>;
-        const classValidationErrors = await validate(instance);
-        const errors = classValidationErrors.map((error) => {
-            const constraints = error.constraints || {};
-            const errorMessage = Object.keys(constraints)
-                .map((constraintKey) => constraints[constraintKey])
-                .join(". ");
-            return { ...error, errorMessage };
-        });
-
-        return { data: instance, errors };
     }
 }
